// CBHIS-Fontend-stable-training-v1.1.1/mobile-apps/src/app/modules/client/enums/hiv.enum.ts
export enum counseling {
  hivInitial = 1,
  hivPostTest = 2,
  anc = 3,
  general = 4,
}

export enum DistributionType {
  SelfTestOnly = 1,
  SelfAndPartnerUse = 2,
  AnotherPerson = 3,
}

export enum UserProfile {
  Pregnant = 1,
  MSM = 2,
  LGBTQ = 3,
  PersonUsingDrugs = 4,
  SexWorker = 5,
  OVC = 6,
}

export enum TestResult {
  Negative = 1,
  Positive = 2,
  NotDisclosed = 3,
}

export enum CounselingType {
  HIVInitial = 1,
  HIVPostTest = 2,
  ANC = 3,
  General = 4,
}